Revert docker files move
These files were initially moved to better support a more modular stack. As
these design decisions are still in flux and may change, revert this change for
now. This keeps the docker instructions simple and doesn't break anyone's
current workflow. This may be worth coming back to and moving again if a more
modular stack is what serves us best in the long run.