Fix options callbacks
In order to support arguments in any order, check for algorithm and parallel
flag compatability with selected environment on the algorithm, env_name, and
parallel options. This ensures that no matter what defaults are used or what
order any argument is passed in all checks are made and any warnings are
displayed and errors are raised.